+++ ExpensApp/composeApp/src/commonMain/kotlin/com/tomasmacri/expensapp/domain/usecases/expense/GetExpenseUseCase.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.flow.flowOf
 
 class GetExpenseUseCase(private val repository: ExpensesRepository) {
 
-    operator fun invoke(id: Int): Flow<Operation<Expense>> {
+    operator fun invoke(id: Long): Flow<Operation<Expense>> {
         if (id <= 0) {
             return flowOf(Operation.Error("The expense does not exist!"))
         }
